# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Documentation for SMIL",
        "contact": {
            "email": "muhammad.raflysadewa.tik17@mhsw.pnj.ac.id"
        },
        "version": "1.0.0"
    },
    "paths": {
        "/api/admin/alat": {
            "get": {
                "tags": [
                    "Alat"
                ],
                "summary": "Return List of Alat",
                "operationId": "getAllAlat",
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "500": {
                        "description": "Internal Server"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Alat"
                ],
                "summary": "Create New Alat",
                "operationId": "createNewAlat",
                "requestBody": {
                    "required": false,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/StoreAlatRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/AlatDetailResource"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "500": {
                        "description": "Internal Server"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/admin/filter/alat": {
            "post": {
                "tags": [
                    "Alat"
                ],
                "summary": "Return List of Alat by Filter",
                "operationId": "getAllAlatByFilter",
                "requestBody": {
                    "required": false,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/FilterAlatRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/AlatListResource"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "500": {
                        "description": "Internal Server"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/admin/alat/{id}": {
            "get": {
                "tags": [
                    "Alat"
                ],
                "summary": "Get Alat Detail Information",
                "operationId": "getAllAlatById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Alat ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/AlatDetailResource"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "500": {
                        "description": "Internal Server"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Alat"
                ],
                "summary": "Update Existed Alat",
                "operationId": "updateExistedAlat",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Alat ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": false,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateAlatRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/AlatDetailResource"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "500": {
                        "description": "Internal Server"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Alat"
                ],
                "summary": "Delete Existed Alat",
                "operationId": "deleteExistingAlat",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Alat ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "500": {
                        "description": "Internal Server"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Page": {
                "title": "Page",
                "description": "Page Model",
                "properties": {
                    "current": {
                        "title": "Current",
                        "description": "Current Page",
                        "type": "integer",
                        "format": "int64",
                        "example": 1
                    },
                    "total": {
                        "title": "Total Page",
                        "description": "Total Page",
                        "type": "integer",
                        "format": "int64",
                        "example": 1
                    },
                    "size": {
                        "title": "Size Page",
                        "description": "Total Data display in Page",
                        "type": "integer",
                        "format": "int64",
                        "example": 1
                    },
                    "data_total": {
                        "title": "Data Total",
                        "description": "Total Data store in system",
                        "type": "integer",
                        "format": "int64",
                        "example": 1
                    }
                },
                "type": "object",
                "xml": {
                    "name": "Page"
                }
            },
...
            "Response": {
                "title": "Response",
                "description": "Response Model",
                "properties": {
                    "code": {
                        "title": "Code",
                        "description": "Code Response",
                        "type": "integer",
                        "format": "int64",
                        "example": 200
                    },
                    "status": {
                        "title": "Status Response",
                        "description": "Status Response",
                        "type": "string",
                        "example": "success"
                    },
                    "message": {
                        "title": "Message Response",
                        "description": "Message Response",
                        "type": "string",
                        "example": "Response Berhasil"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "Response"
                }
            },
...
            "FilterAlatRequest": {
                "title": "Filter Alat request",
                "description": "Filter Alat request body data",
                "properties": {
                    "page_size": {
                        "title": "page_size",
                        "description": "Page Size of Filter",
                        "type": "integer",
                        "example": 5
                    },
                    "sort_by": {
                        "title": "sort_by",
                        "description": "Sort Filter By Attributes",
                        "type": "string",
                        "example": "id"
                    },
                    "sort_direction": {
                        "title": "sort_direction",
                        "description": "Sort Direction of Filter",
                        "type": "string",
                        "example": "ASC"
                    },
                    "alat_name": {
                        "title": "alat_name",
                        "description": "Filter Alat Name",
                        "type": "string"
                    },
                    "jenis_alat_id": {
                        "title": "jenis_alat_id",
                        "description": "Filter Jenis Alat",
                        "type": "integer"
                    },
                    "asal_pengadaan_id": {
                        "title": "asal_pengadaan_id",
                        "description": "Filter Jenis Alat",
                        "type": "integer"
                    },
                    "alat_year": {
                        "title": "alat_year",
                        "description": "Filter Alat Year",
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "StoreAlatRequest": {
                "title": "Store Alat Request",
                "description": "Store Alat Request body data",
                "required": [
                    "alat_name"
                ],
                "properties": {
                    "alat_name": {
                        "title": "alat_name",
                        "description": "Alat Name",
                        "type": "string",
                        "example": ""
                    },
                    "jenis_alat_id": {
                        "title": "jenis_alat_id",
                        "description": "Jenis Alat ID",
                        "type": "integer",
                        "example": 1
                    },
                    "alat_specs": {
                        "title": "alat_specs",
                        "description": "Alat Name",
                        "type": "string"
                    },
                    "asal_pengadaan_id": {
                        "title": "asal_pengadaan_id",
                        "description": "Asal Pengadaan ID",
                        "type": "integer",
                        "example": 1
                    },
                    "alat_year": {
                        "title": "alat_year",
                        "description": "Alat Year",
                        "type": "string"
                    },
                    "supplier_id": {
                        "title": "supplier_id",
                        "description": "Supplier ID",
                        "type": "integer",
                        "example": 1
                    },
                    "alat_total": {
                        "title": "alat_total",
                        "description": "Alat Total",
                        "type": "integer",
                        "example": 10
                    },
                    "lokasi_id": {
                        "title": "lokasi_id",
                        "description": "Lokasi ID",
                        "type": "integer",
                        "example": 1
                    }
                },
                "type": "object"
            },
            "UpdateAlatRequest": {
                "title": "Update Alat Request",
                "description": "Update Alat Request body data",
                "required": [
                    "alat_name"
                ],
                "properties": {
                    "alat_name": {
                        "title": "alat_name",
                        "description": "Alat Name",
                        "type": "string",
                        "example": ""
                    },
                    "jenis_alat_id": {
                        "title": "jenis_alat_id",
                        "description": "Jenis Alat ID",
                        "type": "integer",
                        "example": 1
                    },
                    "alat_specs": {
                        "title": "alat_specs",
                        "description": "Alat Name",
                        "type": "string"
                    },
                    "asal_pengadaan_id": {
                        "title": "asal_pengadaan_id",
                        "description": "Asal Pengadaan ID",
                        "type": "integer",
                        "example": 1
                    },
                    "alat_year": {
                        "title": "alat_year",
                        "description": "Alat Year",
                        "type": "string"
                    },
                    "supplier_id": {
                        "title": "supplier_id",
                        "description": "Supplier ID",
                        "type": "integer",
                        "example": 1
                    }
                },
                "type": "object"
            },
...
            "AlatDetailResource": {
                "title": "AlatDetailResource",
                "description": "Alat Resource",
                "properties": {
                    "response": {
                        "$ref": "#/components/schemas/Response"
                    },
                    "data": {
                        "title": "Data",
                        "description": "Data Alat Wrapper",
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Alat"
                        }
                    }
                },
                "type": "object",
                "xml": {
                    "name": "AlatDetailResource"
                }
            },
            "AlatListResource": {
                "title": "AlatListResource",
                "description": "Alat Resource",
                "properties": {
                    "result": {
                        "title": "Result",
                        "description": "Result Alat Wrapper",
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Alat"
                        }
                    },
                    "page": {
                        "$ref": "#/components/schemas/Page"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "AlatListResource"
                }
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "Authorization",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
    "security": [
        []
    ]
}